requirements.txt
anyio==4.4.0
discord.py==2.3.2
fastapi==0.128.1
matplotlib==3.8.4
numpy==2.4.2
obspy==1.4.2
paho_mqtt==2.1.0
pandas==3.0.0
pydantic==2.12.5
scipy==1.17.0
uvicorn==0.40.0
